package/frr: update to 9.1.1
Update to the latest 9.x version. Protobuf is now mandatory.

The COPYING file has evolved to expose more details: the respective
licenses are now listed, except for the GPLv3 license. The GPLv3
license is only required when FRR is compiled with the GCC plugin that
extends printf type checks. This plugin should not be enabled for
production code. For more information, see FRR's workflow.rst
document.

Signed-off-by: Vincent Jardin <vjardin@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>