package/chrony: install default config
chrony needs a config file to work:
  Could not open /etc/chrony.conf : No such file or directory

Many distros [1][2] use example 2 as the default config, and a few
distros (such as freeBSD [3]) use example 3.

Example 1 [4] is very brief. Example 2 [5] fully includes example 1
and sets the log directory. Example 3 [6] is almost entirely comments
intended to serve as documentation. Therefore, we have chosen to
use example 2 as the out of the box configuration in Buildroot.

[1]: https://github.com/microsoft/azurelinux/blob/3.0-stable/SPECS/chrony/chrony.spec#L72
[2]: https://github.com/vmware/photon/blob/5.0-Beta/SPECS/chrony/chrony.spec#L33
[3]: https://github.com/FreeBSDDesktop/freebsd-ports/blob/master/net/chrony/Makefile#L93-L94
[4]: https://gitlab.com/chrony/chrony/-/blob/3.5-stable/examples/chrony.conf.example1
[5]: https://gitlab.com/chrony/chrony/-/blob/3.5-stable/examples/chrony.conf.example2#L35
[6]: https://gitlab.com/chrony/chrony/-/blob/3.5-stable/examples/chrony.conf.example3#L30-34

Signed-off-by: TIAN Yuanhao <tianyuanhao3@163.com>
Cc: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>