system: change label for BR2_INIT_NONE
Currently, the label reads "None", but it can be a bit misleading.
Indeed, since its inception in 0c00636e8bbc (system/init: add option for
no init system), what it really meant was "none of the above", as
briefly laid out in the commit log for 0c00636e8bbc (typo included):

    Can be usefull if a local package provides an init procedure.

This was later refined in 4d185e5ec51b (system: add help entry to "none"
init system), which explained:

    Add a comment that states the user has to provide his own init
    system, either via a package or a rootfs overlay.

So, change the label to better reflect that.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Reviewed-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>