system: classify init systems
We're going to add more init systems, and some have a very special
purpose, so we do not want to mix them with the mundane platitude of
existing init systems.

Add comments to separate the two.

Note that the list would already contain comments, when a br2-external
tree provides an init system, so those new comments would not be totally
weird-looking.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Reviewed-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>