board/canaan/k210-soc: use tinyinit as Linux init process
The initialization script provided by the tinyinit package is the same
as the one contained in board/canaan/k210-soc/rootfs_overlay/sbin/init
except for the Linux logo, which has been removed. The patch reworks
the configurations that use this overlay directory by replacing the
initialization process contained within it with the one provided by
the tinyinit package.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Reviewed-by: Damien Le Moal <dlemoal@kernel.org>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>