package/lrzsz: fix build with GCC >= 14.x
The lrzsz package currently fails to build with GCC >= 14.x. In [1],
Fabrice had proposed to update the package to what's available on a
Github repository from the upstream maintainer at [2], but in this
very repository (not updated since 4 years), the maintainer writes:

  NOTE:

  Do not use the repository until further notice.
  It contains work in progress and at least one change i am unhappy with.

  Use either http://ohse.de/uwe/testing/lrzsz-0.12.21rc.tar.gz (over
  20 years in rc state, so it's good enough), or 0.12.20 from
  https://ohse.de/uwe/software/lrzsz.html (do not expect me to update
  that old site ever again).

So, let's stick to the 0.12.20 version we have, and add a few patches.

Patch 0001 is patching the configure script, which we obviously don't
normally like. However, the configure.in script is very old and no
longer autoreconfs properly with modern autoconf. Since the patch is
trivial, and upstream seems almost dead, we take the route of patching
the configure script.

[1] https://patchwork.ozlabs.org/project/buildroot/patch/20240726092042.1273023-1-fontaine.fabrice@gmail.com/
[2] https://github.com/UweOhse/lrzsz

Fixes:

  http://autobuild.buildroot.net/results/3eb94eac3e0e1a04d41c618ed9ea2e4e398ea48a/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>