package/popperjs: remove package
PopperJS project is now called floating-ui and is hosted on
https://github.com/floating-ui/floating-ui this caused legacy
source code archives to be repackaged with difference directory
name which broke the sha256 hash and made the autobuild fail.

Given the package has never been bumped since its introduction the
support for that package should probably be dropped.

This package used to be maintained by Thomas De Schampheleire, who no
longer contributes to Buildroot, and removed himself from the
DEVELOPERS file in March 2024.

Fixes:

  https://autobuild.buildroot.org/results/ab62afa6252ad986a900fbfc8b805038edc5d1d6/
  https://autobuild.buildroot.org/results/d1a5b56b987aa178b2ca3fa0855a5bd3da896bcb/
  https://autobuild.buildroot.org/results/9629da94be61a906fd113e77f0120cade521a715/

Signed-off-by: Thomas Perale <thomas.perale@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>