configs/ls1028ardb: update to Linux 6.6
Update Linux, U-Boot, and TF-A to the latest versions used by NXP's
Yocto project.

Drop the patches from board/freescale/common as the TF-A patches are
upstream.

Switch the Linux defconfig to the default ARM64 defconfig, as is done by
the NXP Yocto build.

The updated Linux kernel increases the rootfs size to 122MB so we bump
the rootfs size slightly to match.

See: https://github.com/nxp-qoriq/yocto-sdk/tree/scarthgap

Signed-off-by: Brandon Maier <brandon.maier@collins.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>