package/linux-headers: default to old kernel version
Currently, there is no default value in the choice for the custom kernel
version, and the latest version is at the top of the list.

This means tha a defconfig that uses the latest version does not get the
symbol written in it, and thus when a newer version is latter added, the
defconfig is then incorrect.

To kinda-workaround the issue, make the oldest, pre-3.0 version the
default: this is always going to be the oldest version, we're not going
to add anything older than that, so defconfigs that use that version
will not have it written but that will always yield it back. Conversely,
all other defconfigs will get the kernel version written in them, so it
will stick even when we add newer versions.

Reported-by: Edgar Bonet <bonet@grenoble.cnrs.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>