board/acmesystems/acqua-a5: update device tree for newer kernels
The Ethernet MAC address of the Acqua board is stored in an onboard
EEPROM. Its location is described in the device tree as:

        eeprom@58 {  // NVMEM node
                // EEPROM parameters...

                eth0_addr: eth-mac-addr@9A {  // NVMEM cells
                        reg = <0x0 0x06>;
                };
        };

Since Linux 6.5, defining NVMEM cells this way, as direct children of
the NVMEM node, is deprecated. It is instead recommended to wrap the
cells inside a "fixed-layout" node.

The older, deprecated syntax is still supported on Linux 6.6. Since
Linux 6.7 (commit 2cc3b37f5b6d: "nvmem: add explicit config option to
read old syntax fixed OF cells") however, it is only supported by
selected NVMEM drivers. It is not supported by the at24 driver used to
access the Acqua's onboard EEPROM.

Update the device tree to the new, recommended, syntax. This makes it
work with newer kernels, while preserving compatibility with the
currently used 6.6.30.

Signed-off-by: Edgar Bonet <bonet@grenoble.cnrs.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>