package/coreutils: fix build with uClibc-ng
When compiling coreutils with uClibc-ng, build is failing with error:

    In file included from lib/mcel.c:3:
    lib/mcel.h: In function 'mcel_scan':
    lib/mcel.h:226:21: error: 'mbstate_t' has no member named '__count'
      226 |   mbstate_t mbs; mbs.__count = 0;
          |                     ^

The problem was reported by Waldemar Brodkorb in:
https://lists.gnu.org/r/bug-gnulib/2024-08/msg00130.html

This commit adds a package patch that modifies mcel_scan to avoid
treating uClibc-ng as if it were glibc, resolving compatibility
issues.

Specifically, uClibc-ng was incorrectly treated like glibc in the code,
which led to incorrect assumptions about the mbstate_t initialization.

Fixes:
- https://autobuild.buildroot.org/results/34f4e23e325a03ab37fd5a0a564780b9e0d631cd/

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@yoseli.org>
[Julien:
  - slightly reworded the commit log
  - fixed package patch encoding
  - add the Upstream: tag in package patch
  - add the Fixes: link in commit log
]
Signed-off-by: Julien Olivain <ju.o@free.fr>