package/gnuradio: fix build by removing unneeded package patch
The gnuradio package patch 0001 was introduced in commit 81e6541ef
"package/gnuradio: fix build with fmt >= 11.x" [1] while Buildroot
was in a release client stabilization branch.
It was first in included in the 2024.08-rc2 tag. This is why the
gnuradio version was _not_ bumped at that time.

At the same time, the gnuradio package was also broken in the "next"
branch for the same reason. It was fixed by bumping the gnuradio
version in commit 101e7b69d "package/gnuradio: bump version to
3.10.11.0" [2] which contained the 0001 patch.

Both commits [1] and [2] are individually working. The gnuradio
package fails to build since the merge commit 3daa03752 "Merge branch
'next'" [3].

This commit fixes the issue by removing the package patch.

Fixes:
- http://autobuild.buildroot.org/results/dfda705bb1b3b975bde3d84c6b25a9b41c545c84
- http://autobuild.buildroot.org/results/7c32bd40b5e1420535734480bf5a4d4365b8896f
- https://gitlab.com/buildroot.org/buildroot/-/jobs/7888800399

[1] https://gitlab.com/buildroot.org/buildroot/-/commit/81e6541efe054df635947d0c143818f8cf7cb048
[2] https://gitlab.com/buildroot.org/buildroot/-/commit/101e7b69da88148fa8333af05444e47bae631833
[3] https://gitlab.com/buildroot.org/buildroot/-/commit/3daa037528c38b91caf616cb47992da7730f4343

Signed-off-by: Thomas Perale <thomas.perale@mind.be>
Reviewed-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Julien Olivain <ju.o@free.fr>