package/bluez5_utils: enable asha/bass when building audio plugins
Symbols from the asha plugin are required when building a2dp
support as part of the audio plugins, so always enable asha
when audio plugins are selected.

Symbols from the bass plugin are required when building bap
support as part of the audio plugins, so always enable bass
when audio plugins are selected.

This issue appeared during an attempt to fix a related issue in
ce4e5fd129b557ac0411d0a0704be43717310d6d.

Fixes:
 - http://autobuild.buildroot.net/results/5a3/5a306356f5f190a14a0a2294a4428544b1a018bb
 - http://autobuild.buildroot.net/results/353/3539c83deaa45a3548bcd488659148c62af3fc73

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Cc: Frank TEO <frank128@ymail.com>
Tested-by: Florian Fainelli <f.fainelli@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>