package/numactl: bump to version 2.0.18
For change logs since 2.0.16, see:
https://github.com/numactl/numactl/releases/tag/v2.0.18
https://github.com/numactl/numactl/releases/tag/v2.0.17

This commit also removes the package patch, which was included in
numactl v2.0.17. Since this patch was modifying configure.ac, the
"NUMACTL_AUTORECONF = YES" directive is also removed. Finally, the
patch entry in the ".checkpackageignore" file is also removed.

This commit also adds an upstream patch, not yet in this release, which
fixes builds with Kernel header not providing the
set_mempolicy_home_node() system call.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>