package/acpica: switch to github
As reported in [1] the traditional acpica.org website releases might get
delayed due to website publication process, and indeed, on GitHub, the
latest released version is 20240927, while in [2] the 20240827 version
appears. Furthermore, in [2], explicit reference is made to the GitHub
URL for direct access to the public ACPICA git repository.

[1] https://github.com/acpica/acpica/releases
[2] https://www.intel.com/content/www/us/en/developer/topic-technology/open/acpica/download.html

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>