package/gobject-introspection: don't depend on cairo
Currently, goi depends on cairo.

So far, this would not cause any issue, but once we bump libglib2 and
goi, that would cause a circular dependency chain, where goi would
depend on cairo, cairo on libglib2, and libglib2 back on goi...

However, in goi, cairo is only needed to build some tests, and thus we
do not need that dependency, even before we update goi+libglib2.

Thanks to Fiona for helping sorting the issue out.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Adam Duskett <adam.duskett@amarulasolutions.com>
Cc: Fiona Klute (WIWA) <fiona.klute@gmx.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>