package/arm-gnu-toolchain: use $(HOSTARCH) to allow use on aarch64
When building on an ARM64 build server, the build of
configs/ti_am62x_sk_defconfig fails when building the Cortex-R5
U-Boot, as it uses host-arm-gnu-toolchain, which assumes the host
machine is x86-64.

Use $(HOSTARCH) instead so that it can download the bare-metal
toolchain also provided by ARM for the ARM64 host architecture.

Signed-off-by: Bryce Johnson <bryce@redpinelabs.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>